% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(
  token = NULL,
  redcap_uri = NULL,
  which_report = NULL,
  unique_identifier = NULL
)
}
\arguments{
\item{token}{The user-specific string that serves as the password for a
project.  Required.}

\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.  Required.}

\item{which_report}{A character.
\itemize{
\item \code{All} (the default): creates the full report.
\item \code{recruitment}: only creates the recruitment section of the report
\item \code{safety}: only creates the safety assessment section of the report
}}

\item{unique_identifier}{A character or number. This is the variable name of the
unique identifier in your REDCap project. Commonly the unique id is titled record_id.}
}
\value{
Dataframes
\itemize{
\item \code{df_return}: All relevant data from the REDCap project. This dataframe only contains columns
necessary for the DSM project. Any other columns are removed
\item \code{meta_data}: A dataframe with the project's data dictionary
}
}
\description{
This function uses a REDCap API token to create a labelled data frame. This function heavily relies on
the work from REDCapR. See \link[REDCapR]{redcap_read}
}
\section{Notes}{

20240412: There needs to be a very strict setup in REDCap. AE and violation need
to be repeating instrument. dsmc tool needs to only be in one time point. I need to
check the metadata that the project is setup correctly. Need to build later. JC
20240412: Also need to build checks for the unique identifier. Another note is that I could
rearrage these checks to use the metadata for the project instead of the data. This function
uses redcap_metadata_read(). JC
}

