% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_consort.R
\name{prep_consort}
\alias{prep_consort}
\title{Prep Consort}
\usage{
prep_consort(df = NULL)
}
\arguments{
\item{df}{A dataframe from \code{\link[=get_data]{get_data()}} with all relevant variables}
}
\value{
A labelled dataframe export of project data
}
\description{
This function transforms data from \code{\link[=get_data]{get_data()}} into consort ready
data by sorting participants into groups, i.e. screening, screen-failed, enrolled, etc.
}
